backend/user_service/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
pydantic[email]==2.11.0
pydantic-settings==2.11.0
python-dotenv==1.0.1
passlib==1.7.4
motor==3.7.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.2.0 
aiosmtplib==3.0.1
redis==5.0.8